solution: Fixed project 'Producer-Consumer'
