+++ src/main/java/com/rakovets/course/java/core/practice/concurrency/thread_synchronization/producer_consumer/controller/NumberStoreController.java
@@ -1,14 +1,14 @@
 package com.rakovets.course.java.core.practice.concurrency.thread_synchronization.producer_consumer.controller;
 
-import com.rakovets.course.java.core.practice.concurrency.thread_synchronization.producer_consumer.service.ConsumerThreadService;
-import com.rakovets.course.java.core.practice.concurrency.thread_synchronization.producer_consumer.service.ProducerThreadService;
+import com.rakovets.course.java.core.practice.concurrency.thread_synchronization.producer_consumer.service.ConsumerService;
+import com.rakovets.course.java.core.practice.concurrency.thread_synchronization.producer_consumer.service.ProducerService;
 import com.rakovets.course.java.core.practice.concurrency.thread_synchronization.producer_consumer.service.StoreService;
 
 public class NumberStoreController {
     private final StoreService storeService = new StoreService();
 
     public void start() {
-        new Thread(new ProducerThreadService(storeService)).start();
-        new Thread(new ConsumerThreadService(storeService)).start();
+        new Thread(new ProducerService(storeService)).start();
+        new Thread(new ConsumerService(storeService)).start();
     }
 }
