+++ src/main/java/com/rakovets/course/java/core/practice/concurrency/thread_synchronization/producer_consumer/controller/NumberStoreControllerStarter.java
@@ -1,6 +1,6 @@
 package com.rakovets.course.java.core.practice.concurrency.thread_synchronization.producer_consumer.controller;
 
-public class NumberStoreStarterController {
+public class NumberStoreControllerStarter {
     public static void main(String[] args) {
         NumberStoreController numberStoreController = new NumberStoreController();
         numberStoreController.start();
