+++ src/main/java/com/rakovets/course/java/core/practice/concurrency/thread_synchronization/producer_consumer/service/ConsumerService.java
@@ -1,23 +1,25 @@
 package com.rakovets.course.java.core.practice.concurrency.thread_synchronization.producer_consumer.service;
 
-public class ProducerThreadService implements Runnable {
-    private final StoreService store;
-    protected boolean isRun = true;
+import java.util.Random;
 
-    public ProducerThreadService(StoreService store) {
+public class ConsumerService implements Runnable {
+    private final StoreService store;
+    public boolean isRun = true;
+
+    public ConsumerService(StoreService store) {
         this.store = store;
     }
 
     public void run() {
         while (isRun) {
-            store.produce();
-            sleep();
+            sleep(new Random().nextInt(1000));
+            store.consume();
         }
     }
 
-    private static void sleep() {
+    private static void sleep(long millis) {
         try {
-            Thread.sleep(3000);
+            Thread.sleep(millis);
         } catch (InterruptedException e) {
             e.printStackTrace();
         }
